cypress: BLN: Add led_blink support
This enables led_blink support. When a blinking configuration is set,
the led_blink will use a wakelock to handle the blinking-effect. After
the default led_blink_timeout of 10 minutes, the blinking will stop,
the LED is lit and the wakelock is released.

To enable led_blink, write 1 to
/sys/devices/virtual/misc/backlightnotification/led_blink_enabled

The led_blink timeout can be set by writing milliseconds to
/sys/devices/virtual/misc/backlightnotification/led_blink_timeout

Change-Id: I272451c1812b0affe2c1b170010a57497f3ec3da